requirements.txt

flask
flask-cors
pymysql
python-dotenv
requests

PyPDF2
sentence-transformers
numpy
torch
transformers